Simplified making of libcrunky
vectors.o is now part of libcrunky. There was a whole complication
where the vectors didn't seem to put into final elf file. Now
they do.

I'm not sure that the matter has been completely put to bed,
but it seems to be working now. Just keep an eye on it,
just in case.