Improved flashing
It sure as hell isn't making things easy!

It appears that you can't just do a bulk erase. You have to erase,
then write to the page. So it seems as if we need to erase a page
(4k) and write a page (4k). I'm not sure if that's strictly
necessary, but there seem to be some odd restrictions.